+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - improve performance when tracking a big memory dump (~1500 bytes).
 - only lookup for properties within boundaries of memory dump.
+- only show hovered state when mouse is inside of MEMORY / PROPERTIES views.
 
 ------------------------