+++ CHANGELOG.md
@@ -15,7 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - support to print data types Quaternion, EulerAngles, WorldPosition and 
   WorldTransform.
 - scroll to property's offset in MEMORY view when hovering it in PROPERTIES 
-  view.
+  view. No scroll when a property is selected.
 
 ------------------------