+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+------------------------
+
+## [0.2.0] - 2024-05-08
 ### Fixed
 - improve performance when tracking a big memory dump (~1500 bytes).
 - only lookup for properties within boundaries of memory dump.
@@ -12,9 +16,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - offset selection in DATA VIEWER.
 
 ### Added
-- support to print data types Quaternion, EulerAngles, WorldPosition and 
+- support to print data types Quaternion, EulerAngles, WorldPosition and
   WorldTransform.
-- scroll to property's offset in MEMORY view when hovering it in PROPERTIES 
+- scroll to property's offset in MEMORY view when hovering it in PROPERTIES
   view. No scroll when a property is selected.
 
 ------------------------
@@ -26,6 +30,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - CET api to allow other mods to track targets.
 
 <!-- Table of releases -->
-[Unreleased]: https://github.com/rayshader/cp2077-red-memorydump/compare/v0.1.0...HEAD
+[Unreleased]: https://github.com/rayshader/cp2077-red-memorydump/compare/v0.2.0...HEAD
 [0.2.0]: https://github.com/rayshader/cp2077-red-memorydump/compare/v0.1.0...v0.2.0
 [0.1.0]: https://github.com/rayshader/cp2077-red-memorydump/releases/tag/v0.1.0